Backend/requirements.txt
flask==3.1.3
flask-cors==6.0.2
python-dotenv==1.2.2
pandas==2.3.3
scikit-learn==1.8.0
openai==2.26.0
google-generativeai==0.8.6
langchain==1.2.12
langchain-core==1.2.18
langchain-google-genai==4.2.1
langchain-openai==1.1.11
langgraph==1.1.2
